Se programo la clase del Registro Devoluciones y se cambio el textBox de EstudianteId por un ComboBox
Se cambiaron los tipos de datos de algunas variables de las clases Devoluciones y DevolucionesDetalle

+++ Entidades/DevolucionesDetalle.cs
@@ -14,8 +14,8 @@ namespace ProyectoFinal_PrestamosLibros.Entidades
         public int DevolucionDetalleId { get; set; }
         public int DevolucionId { get; set; }
         public int LibroId { get; set; }
-        public int LibrosDevueltos { get; set; }
-        public int Dias { get; set; }
+        public double LibrosDevueltos { get; set; }
+        public double Dias { get; set; }
 
         //———————————————————————————[ ForeingKey ]———————————————————————————
         [ForeignKey("LibroId")]
